Refactor save_result
